Add files via upload
Master’s Thesis :- Cubic B-Spline Collocation Method for Parabolic Partial Differential Equations.
Project descriptionThe aim of this Dissertation is to investigate the features of the numerical solution of parabolic partial differential equations(one-dimensional heat Equations). In this thesis, we saw the advantage of B-spline Collocation method in numerical Process. Which is provide a high order of accuracy to understand in many physical Process.